@@ -5,7 +5,7 @@ description: A new Flutter application.
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">=2.10.0 <3.0.0"
 
 dependencies:
   flutter:
@@ -15,6 +15,7 @@ dependencies:
 
   cupertino_icons: ^0.1.3
   validators: ^2.0.1
+  shared_preferences: ^0.5.12+2
 
 dev_dependencies:
   flutter_test:
@@ -31,5 +32,6 @@ flutter:
       - family: Nunito
         fonts:
           - asset: assets/fonts/Nunito-Regular.ttf
+
 
 
